Solved the issue of bottom sheet in Logic Analyzer instrument
Added private level sccoping
