--- app/src/main/res/layout-port/logic_analyzer_main_layout.xml
@@ -1,17 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<android.support.design.widget.CoordinatorLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent">
-
-    <include layout="@layout/logic_analyzer_logic_lines" />
-
-    <View
-        android:id="@+id/shadow"
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        android:alpha="0"
-        android:background="@color/black" />
-
-    <include layout="@layout/bottom_sheet_custom" />
-
-</android.support.design.widget.CoordinatorLayout>
