+++ app/src/main/res/layout/bottom_sheet_custom.xml
@@ -13,11 +13,7 @@
 
     <Space
         android:layout_width="match_parent"
-        android:layout_height="?attr/actionBarSize" />
-
-    <Space
-        android:layout_width="match_parent"
-        android:layout_height="?attr/actionBarSize" />
+        android:layout_height="@dimen/extra_space_top" />
 
     <LinearLayout
         android:layout_width="match_parent"
@@ -59,6 +55,10 @@
                 android:orientation="vertical"
                 android:padding="@dimen/bottom_sheet_padding">
 
+                <Space
+                    android:layout_width="match_parent"
+                    android:layout_height="@dimen/space_length_btm_sheet" />
+
                 <TextView
                     android:id="@+id/guide_title"
                     android:layout_width="match_parent"
@@ -92,6 +92,10 @@
                     android:layout_width="match_parent"
                     android:layout_height="wrap_content"
                     android:orientation="horizontal" />
+
+                <Space
+                    android:layout_width="match_parent"
+                    android:layout_height="@dimen/space_length_btm_sheet" />
 
             </LinearLayout>
